[localization] Modifies synchronization algorithm (just skeleton here).

+++ static/localization/balise_principal/lib/lib_domitille.cpp
@@ -9,7 +9,7 @@ int _pinInt;
 
 //number of revolution
 volatile unsigned int _nbTR=0;
-volatile unsigned long prev_int=0,last_TR=0,TR_period=0,TR_mean_period=0;
+volatile uint32_t prev_int=0,last_TR=0,TR_period=0,TR_mean_period=0;
 #define FILTER_SHIFT 2
 volatile unsigned long filter_reg=0;
